# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentCreate": {
        "additionalProperties": false,
        "properties": {
          "agreed": {
            "title": "Agreed",
            "type": "boolean"
          },
          "consentVersion": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Consentversion",
            "type": "string"
          }
        },
        "required": [
          "consentVersion",
          "agreed"
        ],
        "title": "ConsentCreate",
        "type": "object"
      },
      "ConsentRead": {
        "additionalProperties": false,
        "properties": {
          "agreed": {
            "title": "Agreed",
            "type": "boolean"
          },
          "agreedAt": {
            "format": "date-time",
            "title": "Agreedat",
            "type": "string"
          },
          "consentVersion": {
            "title": "Consentversion",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "consentVersion",
          "agreed",
          "agreedAt"
        ],
        "title": "ConsentRead",
        "type": "object"
      },
...
      "DeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deletedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletedat"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
      "DepartmentListRead": {
        "additionalProperties": false,
        "properties": {
          "departments": {
            "items": {
              "$ref": "#/components/schemas/DepartmentRead"
            },
            "title": "Departments",
            "type": "array"
          }
        },
        "required": [
          "departments"
        ],
        "title": "DepartmentListRead",
        "type": "object"
      },
      "DepartmentRead": {
        "additionalProperties": false,
        "properties": {
          "college": {
            "title": "College",
            "type": "string"
          },
          "curriculumUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculumurl"
          },
          "graduationUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduationurl"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "college",
          "name",
          "curriculumUrl",
          "graduationUrl"
        ],
        "title": "DepartmentRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserDeleteRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "UserDeleteRequest",
        "type": "object"
      },
      "UserRead": {
        "additionalProperties": false,
        "properties": {
          "admissionYear": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admissionyear"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "entryType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrytype"
          },
          "grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "majorPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Majorpath"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "profileCompleted": {
            "title": "Profilecompleted",
            "type": "boolean"
          },
          "sectionGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sectiongroup"
          },
          "studentNumber": {
            "title": "Studentnumber",
            "type": "string"
          },
          "studentType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studenttype"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "studentNumber",
          "name",
          "grade",
          "department",
          "admissionYear",
          "entryType",
          "studentType",
          "sectionGroup",
          "majorPath",
          "profileCompleted",
          "createdAt",
          "updatedAt"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "UserUpdate": {
        "additionalProperties": false,
        "properties": {
          "admissionYear": {
            "anyOf": [
              {
                "maximum": 2100.0,
                "minimum": 1990.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admissionyear"
          },
          "department": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "entryType": {
            "anyOf": [
              {
                "enum": [
                  "FRESHMAN",
                  "TRANSFER"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrytype"
          },
          "grade": {
            "anyOf": [
              {
                "maximum": 4.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "majorPath": {
            "anyOf": [
              {
                "enum": [
                  "ADVANCED_MAJOR",
                  "DOUBLE_MAJOR",
                  "MINOR",
                  "MICRO_MAJOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Majorpath"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sectionGroup": {
            "anyOf": [
              {
                "enum": [
                  "ODD",
                  "EVEN",
                  "UNKNOWN"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sectiongroup"
          },
          "studentType": {
            "anyOf": [
              {
                "enum": [
                  "DOMESTIC",
                  "INTERNATIONAL",
                  "UNKNOWN"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studenttype"
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PL-timeTabler API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/departments": {
      "get": {
        "operationId": "list_departments_api_v1_departments_get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Departments",
        "tags": [
          "departments"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "delete": {
        "operationId": "delete_me_api_v1_users_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Me",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "update_me_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Me",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me/consents": {
      "get": {
        "operationId": "list_consents_api_v1_users_me_consents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentRead"
                  },
                  "title": "Response List Consents Api V1 Users Me Consents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Consents",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_consent_api_v1_users_me_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Consent",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}